+++ src/Modules/Blog/BlogTransactions.php
@@ -3,8 +3,10 @@
 namespace Foodsharing\Modules\Blog;
 
 use Foodsharing\Lib\Session;
+use Foodsharing\Modules\Blog\DTO\BlogPost;
 use Foodsharing\Modules\Core\DBConstants\Uploads\UploadUsage;
 use Foodsharing\Modules\Uploads\UploadsGateway;
+use Foodsharing\Modules\Uploads\UploadsTransactions;
 use Foodsharing\RestApi\Models\Blog\BlogPostData;
 
 class BlogTransactions
@@ -12,6 +14,7 @@ class BlogTransactions
     public function __construct(
         private readonly BlogGateway $blogGateway,
         private readonly UploadsGateway $uploadsGateway,
+        private readonly UploadsTransactions $uploadsTransactions,
         private readonly Session $session,
     ) {
     }
@@ -41,6 +44,19 @@ class BlogTransactions
             // cut the `/api/uploads/` in front of the UUID
             $uuid = substr($post->picture, 13);
             $this->uploadsGateway->setUsage([$uuid], UploadUsage::BLOG_POST, $post->id);
+        }
+    }
+
+    /**
+     * Permanently delets a blog post. This also deletes the post's picture, if it has any.
+     */
+    public function deleteBlogPost(BlogPost $post): void
+    {
+        $this->blogGateway->del_blog_entry($post->id);
+
+        if (!empty($post->picture)) {
+            $oldUUID = substr($post->picture, 13);
+            $this->uploadsTransactions->deleteUploadedFile($oldUUID);
         }
     }
 }