+++ src/Modules/FoodSharePoint/FoodSharePointTransactions.php
@@ -124,4 +124,16 @@ class FoodSharePointTransactions
 
         $this->foodSharePointGateway->updateFSPManagers($currentData['id'], $newData->managerIds);
     }
+
+    public function deleteFoodSharePoint(int $foodSharePointId): void
+    {
+        // Delete the food share point's title picture
+        $foodSharePoint = $this->foodSharePointGateway->getFoodSharePoint($foodSharePointId);
+        if (!empty($foodSharePoint['picture'])) {
+            $uuid = substr($foodSharePoint['picture'], 13);
+            $this->uploadsTransactions->deleteUploadedFile($uuid);
+        }
+
+        $this->foodSharePointGateway->deleteFoodSharePoint($foodSharePointId);
+    }
 }