+++ CHANGELOG.md
@@ -95,6 +95,7 @@
 - Added progress bars to the seed script !3663 @SaiCode @alex.simm
 - Fix tests in ForumPostCest !3998 @chriswalg
 - Added retry and more timeout in WorkingGroupApiCest->sendMailToGroup !4007 @chriswalg
+- Delete uploaded pictures when the corresponding entity is deleted #1182 !4081 @alex.simm
 
 # Hotfix 2024-01-20